requirements.txt
asgiref==3.6.0
certifi==2019.9.11
cffi==1.15.1
chardet==3.0.4
cryptography==39.0.2
defusedxml==0.6.0
Django==4.1.7
django-allauth==0.52.0
django-crispy-forms==2.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
idna==2.8
oauthlib==3.1.0
phonenumbers==8.13.7
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.5
